+++ Changelog.md
@@ -1,3 +1,6 @@
+### ToyBox Rogue - Ver 2.0.13 (built for 1.6.1.511)
+* (***ADDB***) Re-add Bag of Tricks => Other Multipliers => Vision Range Multiplier and make it persist across restarts and area loads.
+
 ### ToyBox Rogue - Ver 2.0.12 (built for 1.6.1.511)
 * (***ADDB***) Bag of Tricks => Remote Companion Dialog: Try to fix issue of Voice Overs not being played by remote units? (Can someone please report whether this is still an issue?)
 * (***ADDB***) Party => Stats => Skeleton Replacer: